Fixed how Solana transaction are indexed when using program + account
The index now uses the full account maps instead of just those from the instructions.

Furthermore, the way the filtering is performed was scoped to instruction previously, e.g. that a full instruction needed to match the query to consider a transaction included. This is now changed, we instead take the set of program ids and accounts defined by the transaction, according to the new rules for accounts above, and if there is a match between this set, then the transaction is included.

This in most cases shouldn't affect anyone, as downstream code usually further filter instructions they are interesting in. But this changes is still consider breaking, as with the new methodology, transactions that didn't match before will now match.